CS 2420 (Data Structures & Algorithms)/assign10/FindKLargest.java
package assign10;

import java.util.ArrayList;
import java.util.Collections;
import java.util.Comparator;
import java.util.List;

/**
 * This class contains generic static methods for finding the k largest items in
 * a list.
 * 
 * @author Erin Parker, Isaac Gibson, and Matthew Johnsen
 * @version April 8, 2019
 */
public class FindKLargest {

	/**
	 * Determines the k largest items in the given list, using a binary max heap and
	 * the natural ordering of the items.
	 * 
	 * @param items - the given list
	 * @param k     - the number of largest items
	 * @return a list of the k largest items, in descending order
	 * @throws IllegalArgumentException if k is negative or larger than the size of
	 *                                  the given list
	 */
	public static <E extends Comparable<? super E>> List<E> findKLargestHeap(List<E> items, int k)
			throws IllegalArgumentException {
		ArrayList<E> list = new ArrayList<>();
		BinaryMaxHeap<E> heap = new BinaryMaxHeap<>(items);
		for (int i = 0; i < k; i++) {
			list.add(heap.extractMax());
		}
		return list;
	}

	/**
	 * Determines the k largest items in the given list, using a binary max heap.
	 * 
	 * @param items - the given list
	 * @param k     - the number of largest items
	 * @param cmp   - the comparator defining how to compare items
	 * @return a list of the k largest items, in descending order
	 * @throws IllegalArgumentException if k is negative or larger than the size of
	 *                                  the given list
	 */
	public static <E> List<E> findKLargestHeap(List<E> items, int k, Comparator<? super E> cmp)
			throws IllegalArgumentException {
		ArrayList<E> list = new ArrayList<>();
		BinaryMaxHeap<E> heap = new BinaryMaxHeap<>(items, cmp);
		for (int i = 0; i < k; i++) {
			list.add(heap.extractMax());
		}
		return list;
	}

	/**
	 * Determines the k largest items in the given list, using Java's sort routine
	 * and the natural ordering of the items.
	 * 
	 * @param items - the given list
	 * @param k     - the number of largest items
	 * @return a list of the k largest items, in descending order
	 * @throws IllegalArgumentException if k is negative or larger than the size of
	 *                                  the given list
	 */
	public static <E extends Comparable<? super E>> List<E> findKLargestSort(List<E> items, int k)
			throws IllegalArgumentException {
		Collections.sort(items);
		ArrayList<E> list = new ArrayList<>();
		for (int i = 1; i <= k; i++) {
			list.add(items.get(items.size() - i));
		}
		return list;
	}

	/**
	 * Determines the k largest items in the given list, using Java's sort routine.
	 * 
	 * @param items - the given list
	 * @param k     - the number of largest items
	 * @param cmp   - the comparator defining how to compare items
	 * @return a list of the k largest items, in descending order
	 * @throws IllegalArgumentException if k is negative or larger than the size of
	 *                                  the given list
	 */
	public static <E> List<E> findKLargestSort(List<E> items, int k, Comparator<? super E> cmp)
			throws IllegalArgumentException {
		Collections.sort(items, cmp);
		ArrayList<E> list = new ArrayList<>();
		for (int i = 1; i <= k; i++) {
			list.add(items.get(items.size() - i));
		}
		return list;
	}
}